@@ -1,24 +1,20 @@
-#include <QVulkanWindow>
-#include <QVulkanFunctions>
-
 #include "MappedData.h"
+#include "Common.h"
 #include "Texture.h"
 #include "Utils.h"
 
 namespace Vulkan
 {
-    MappedData::MappedData(const Image& image, const QVulkanWindow& window)
-        : m_window(window)
+    MappedData::MappedData(const Image& image, const VulkanShared& vulkan)
+        : vulkan(vulkan)
         , memory(image.memory)
     {
-        auto& device_functions = *window.vulkanInstance()->deviceFunctions(window.device());
-
         VkImageSubresource subres = { VK_IMAGE_ASPECT_COLOR_BIT, 0, 0 };
         VkSubresourceLayout layout;
-        device_functions.vkGetImageSubresourceLayout(window.device(), image.image, &subres, &layout);
+        vkGetImageSubresourceLayout(vulkan.device, image.image, &subres, &layout);
 
-        VkResultSuccess(device_functions.vkMapMemory(
-            window.device(), image.memory, layout.offset, layout.size, 0, reinterpret_cast<void**>(&data)
+        VkResultSuccess(vkMapMemory(
+            vulkan.device, image.memory, layout.offset, layout.size, 0, reinterpret_cast<void**>(&data)
         ));
     }
 
@@ -27,7 +23,7 @@ namespace Vulkan
         if (!memory)
             return;
 
-        m_window.vulkanInstance()->deviceFunctions(m_window.device())->vkUnmapMemory(m_window.device(), memory);
+        vkUnmapMemory(vulkan.device, memory);
     }
 
     uint8_t*& MappedData::GetData()
