@@ -2,23 +2,22 @@
 
 #include <vulkan/vulkan.h>
 
-class QVulkanWindow;
-
 namespace Vulkan
 {
 
+struct VulkanShared;
 struct Image;
 
 class MappedData
 {
 public:
-    MappedData(const Image& image, const QVulkanWindow& window);
+    MappedData(const Image& image, const VulkanShared& vulkan);
     ~MappedData();
 
     uint8_t*& GetData();
 
 private:
-    const QVulkanWindow& m_window;
+    const VulkanShared& vulkan;
 
     VkDeviceMemory memory = nullptr;
     uint8_t*       data = nullptr;
