@@ -3,11 +3,10 @@
 #include <vulkan/vulkan.h>
 #include "IRenderer.h"
 
-class QVulkanDeviceFunctions;
-class QVulkanWindow;
-
 namespace Vulkan
 {
+
+struct VulkanShared;
 
 struct Image
 {
@@ -19,14 +18,13 @@ class Texture
     : public ITexture
 {
 public:
-    Texture(const IDataProvider& data, const QVulkanWindow& window, VkFormat format = VK_FORMAT_R8G8B8A8_UNORM);
+    Texture(const IDataProvider& data, VulkanShared& vulkan, VkFormat format = VK_FORMAT_R8G8B8A8_UNORM);
     ~Texture() override;
 
     VkDescriptorImageInfo GetInfo() const;
 
 private:
-    QVulkanDeviceFunctions& functions;
-    VkDevice device = nullptr;
+    VulkanShared& vulkan;
 
     Image image{};
     VkSampler      sampler = nullptr;
